Functionality Improvements and Advancements.
Improved the functionality of some of the move declaration functions
for parsing input and determining available moves.

Began implementation of the functionality for taking move input then
returning a list of the tiles that should be flipped.

TODO: Need to suss out how to create the list of items I need so I can
abstract it out. :(