Updated Clojure, Added Core.Match, Game Updates.
Updated to Clojure 1.5.1, this should be awesome... in theory.

Added core.match to simplify some functions, not much use yet but
this will improve.

Started really hacking on the game loop to actually make this sucker
playable. Only the first move is working, the entire game dies when
the second move is made. Not entirely sure why this is happening since
the functions involved seem to checkout okay.. At least in inital
testing.

Probably should start hacking on a proper suite of tests for this to
make sure I'm not doing something dense after the state of the game
updates.

Bleh...