Refined the Move Creation functionality.
Jiggying with the move sequence functionality to make it work properly.
Currently it is broken when dealing with sequences like :W :B :W :B, as
a :B player this will result in:

This
:W :B :W :B

Becoming this
:B :B W: :B

Which if memory serves is wrong as that final :W should also flipped
because the entire sequence is capped with :B, thus creating a full set
of tiles to flip.

Most of the moves will work okay except that bug kinda ruins the game,
so I've some fail to fix up there.

Targets for fix are:

move-sequence-valid?
create-board-adjustment
get-flippable-tiles
and
squares-to-flip

There is too many moving parts in these functions and I need to break it
down into more sensible steps. Bad code smell at the moment...